safari-600.7-branch/LayoutTests/fast/url/host-lowercase-per-scheme-expected.txt

Canonical URL should have a lowercase hostname for server based protocols.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
FAIL src should be ghost://UnicodeF%C3%AAte/. Was ghost://xn--unicodefte-t7a/.
PASS successfullyParsed is true

TEST COMPLETE
